# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: API - Aplicação de Planejamento de Viagens em Grupo
  description: |
    Documentação da API para organização colaborativa de viagens em grupo.
    Contém endpoints para autenticação, gerenciamento de grupos, destinos,
    votações, despesas e tarefas.
  version: 1.0.0
  contact:
    name: Felipe Lopes Firmino
    email: felipe@example.com
servers:
  - url: http://localhost:8080
    description: Servidor local
  - url: https://api.viagens.com
    description: Servidor de produção
...
paths:
  /auth/login:
    post:
      tags: [Autenticação]
      summary: Realiza login de um usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "401":
          description: Credenciais inválidas
  /auth/register:
    post:
      tags: [Autenticação]
      summary: Cadastra um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        "201":
          description: Usuário criado com sucesso
        "409":
          description: Conflito (E-mail já está em uso)
...
  /groups:
    post:
      tags: [Grupos de Viagem]
      summary: Cria um novo grupo de viagem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TravelGroupCreateRequest'
      responses:
        "201":
          description: Grupo criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelGroupResponse'
        "401":
          description: Não autorizado (Token ausente ou inválido)
        "422":
          description: 'Entidade não processável (Dados inválidos, ex: datas erradas)'
    get:
      tags: [Grupos de Viagem]
      summary: Lista grupos do usuário logado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TravelGroupListItem'
        "401":
          description: Não autorizado (Token ausente ou inválido)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
          example: usuario@email.com
        password:
          type: string
          example: senha123
    UserRegisterRequest:
      type: object
      properties:
        name:
          type: string
          example: João Silva
        email:
          type: string
          example: usuario@email.com
        password:
          type: string
          example: senha123
...
    TravelGroupCreateRequest:
      type: object
      required:
        - name
        - start_date
        - end_date
      properties:
        name:
          type: string
          example: Viagem para Rio de Janeiro
        description:
          type: string
          description: Uma breve descrição ou tema da viagem.
          example: Uma viagem focada em praias e pontos turísticos clássicos.
        start_date:
          type: string
          format: date
          example: 2025-12-10
        end_date:
          type: string
          format: date
          example: 2025-12-20
    TravelGroupResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID gerado pelo banco de dados.
          example: 10
        name:
          type: string
        description:
          type: string
          nullable: true
        creator_id:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    TravelGroupListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          description: Descrição do grupo (para exibição em cards/listas).
          nullable: true
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        member_count:
          type: integer
        creator_id:
          type: integer
        creator_name:
          type: string
...